requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
lxml==4.6.3
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
packaging==20.9
pandas==1.1.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
